pyproject.toml
[project]
name = "document_analyser"
version = "0.1.0"
description = ""
authors = [
    {name = "Joao"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pypdf (>=5.4.0,<6.0.0)",
    "langchain-core (>=0.3.59,<0.4.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "langchain-postgres (>=0.0.14,<0.0.15)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "langchain-community (<0.3.24)",
]
packages = [{ include = "document_analyser", from = "src" }]

[tool.hatch.build.targets.wheel]
packages = ["src/document_analyser"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"